+++ migrations/2025-04-17-160808_create_attendance/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE attendance (
     student TEXT NOT NULL,
-    week INTEGER NOT NULL,
+    week BIGINT NOT NULL,
     status TEXT NOT NULL,
     FOREIGN KEY (student) REFERENCES students (id) ON DELETE CASCADE,
     FOREIGN KEY (week) REFERENCES weeks (id) ON DELETE CASCADE,
